Renaming index.ts to database-handler.ts
